[hotfix][runtime] Consolidating javadoc and package structure.
